build(cmake): add compiler hardening, sanitizers, git hooks, and formatting targets
Integrate comprehensive security hardening flags, compiler sanitizers,
automated Git pre-commit formatting hooks, and clang-format CMake targets
across all project libraries, executables, and tests.

- cmake/CompilerFlags.cmake:
  * Provide configurable options for WARNINGS_AS_ERRORS, ENABLE_HARDENING,
    ENABLE_ASAN, ENABLE_UBSAN, and ENABLE_TSAN.
  * GCC/Clang: Apply strict diagnostics (-Wall, -Wextra, -Wpedantic, -Wshadow,
    -Wconversion, -Wsign-conversion, -Wnull-dereference, -Wdouble-promotion).
  * Security hardening (ENABLE_HARDENING):
    - Stack protection: -fstack-protector-strong and -fstack-clash-protection.
    - Control-flow integrity: -fcf-protection=full on x86_64.
    - ASLR / PIE: -fPIE and -pie on executables.
    - Linker security: -Wl,-z,relro,-z,now and -Wl,-z,noexecstack.
    - Source fortification: -U_FORTIFY_SOURCE -D_FORTIFY_SOURCE=2 in optimized builds.
  * MSVC: Configure /W4, /WX, /GS, /guard:cf, /NXCOMPAT, /DYNAMICBASE,
    /HIGHENTROPYVA, and /fsanitize=address.
  * Apply flags systematically via apply_compiler_flags() across FujinonSX800Core,
    FujinonSX800Qt, FujinonSX800App, and all test targets.

- cmake/GitHooks.cmake & cmake/pre-commit:
  * Automatically install and configure executable git pre-commit hook into
    ${GIT_HOOKS_DIR}/pre-commit using `git rev-parse --git-path hooks`.
  * Add custom targets:
    - install-git-hooks: Explicitly install/update pre-commit hook.
    - pre-commit: Directly invoke pre-commit hook script.
    - format: Reformat all repository C++ sources with clang-format.
    - check-format: Validate formatting via clang-format --dry-run --Werror.

- Source fixes & formatting:
  * Format all C++ source and header files to comply with .clang-format.
  * Fix sign-conversion warning with socklen_t in TcpTransport::open().
  * Fix qsizetype to std::size_t conversion in TrafficInspectorWidget::logFrame().

+++ app-qt/widgets/TrafficInspectorWidget.cpp
@@ -95,7 +95,7 @@ void TrafficInspectorWidget::logFrame(const QByteArray& frame, bool isTx)
     const QString timeStr = QDateTime::currentDateTime().toString("hh:mm:ss.zzz");
     const QString dirStr = isTx ? "TX  >>" : "RX  <<";
 
-    std::vector<std::uint8_t> stdBytes(frame.size());
+    std::vector<std::uint8_t> stdBytes(static_cast<std::size_t>(frame.size()));
     for (qsizetype i { 0 }; i < frame.size(); ++i) {
         stdBytes[static_cast<std::size_t>(i)] = static_cast<std::uint8_t>(frame.at(i));
     }
